+++ src/components/Canvas/context.ts
@@ -1,4 +1,7 @@
 import { TldrawApp } from "@tldraw/tldraw";
 import React from "react";
 
+/**
+ * Provides access to the TLDraw instance.
+ */
 export const AppContext = React.createContext<TldrawApp>({} as any);