+++ src/types.ts
@@ -11,10 +11,26 @@ export interface FSBoard {
   bindings: string[];
 }
 
+export type BoardId = string;
+export type UserId = string;
+
+export type BoardMeta = {
+  id: BoardId;
+  createdBy: UserId;
+  createdOn: Date;
+};
+
+export type FSUser = {
+  id: UserId;
+};
+
 export interface FSAdapter {
   isLoading: boolean;
   createDocument: () => string;
   loadDocument: (input: Uint8Array) => void;
   serialiseDocument: () => Uint8Array;
   eventHandlers: TldrawProps;
+  createDuplicate: (boardId: BoardId) => BoardId;
+  board: BoardMeta;
+  user: FSUser;
 }