+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use stitches for CSS-in-JS [#56](https://github.com/interalia-studio/fullscreen/pull/56)
 - Self-hosted Urbanist web font [#59](https://github.com/interalia-studio/fullscreen/pull/59)
 - More tools for the new toolbar [#61](https://github.com/interalia-studio/fullscreen/pull/61)
+- Onboarding dialogue when joining a board on the web [#66](https://github.com/interalia-studio/fullscreen/pull/66)
 
 ### Changed